Solucionando errores.

+++ TFG-ClubAppManager/src/main/java/com/umu/springboot/servicio/ServicioFotos.java
@@ -1,10 +1,14 @@
 package com.umu.springboot.servicio;
 
+import java.awt.image.BufferedImage;
+import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
+
+import javax.imageio.ImageIO;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -13,6 +17,7 @@ import org.springframework.web.multipart.MultipartFile;
 
 import com.umu.springboot.modelo.Imagen;
 import com.umu.springboot.repositorios.RepositorioImagen;
+import com.umu.springboot.utils.WebPUtilidades;
 
 @Service
 @Transactional
@@ -20,6 +25,9 @@ public class ServicioFotos implements IServicioFotos {
 
 	@Autowired
 	private RepositorioImagen repositorioImagen;
+	
+	@Autowired
+	private WebPUtilidades webPConverter;
 
 	@Override
 	public List<Long> almacenarFotos(MultipartFile dniFrontal, MultipartFile dniTrasero, MultipartFile dniFrontalTutor1,
@@ -41,20 +49,6 @@ public class ServicioFotos implements IServicioFotos {
 				.map(Imagen::getId).collect(Collectors.toList());
 	}
 
-	private boolean esArchivoValido(MultipartFile file) {
-		return file != null && !file.isEmpty();
-	}
-
-	private Imagen crearImagen(MultipartFile file) {
-		Imagen img = null;
-		try {
-			img = new Imagen(file.getOriginalFilename(), file.getBytes());
-		} catch (IOException e) {
-			e.printStackTrace();
-		}
-		return img;
-	}
-
 	@Override
 	public List<Imagen> descargarFotos(long dniFrontal, long dniTrasero, long dniFrontalTutor1, long dniTraseroTutor1,
 			long dniFrontalTutor2, long dniTraseroTutor2) {
@@ -62,7 +56,7 @@ public class ServicioFotos implements IServicioFotos {
 		List<Long> archivos = Arrays.asList(dniFrontal, dniTrasero, dniFrontalTutor1, dniTraseroTutor1,
 				dniFrontalTutor2, dniTraseroTutor2);
 		
-		return archivos.stream().map(repositorioImagen::findById).filter(Optional::isPresent).map(Optional::get).collect(Collectors.toList());		
+		return archivos.stream().map(repositorioImagen::findById).filter(Optional::isPresent).map(Optional::get).map(this::recuperarImagen).collect(Collectors.toList());		
 	}
 	
 	@Override
@@ -70,7 +64,7 @@ public class ServicioFotos implements IServicioFotos {
 		
 		List<Long> archivos = Arrays.asList(dniFrontal, dniTrasero, certDelitos);
 		
-		return archivos.stream().map(repositorioImagen::findById).filter(Optional::isPresent).map(Optional::get).collect(Collectors.toList());		
+		return archivos.stream().map(repositorioImagen::findById).filter(Optional::isPresent).map(Optional::get).map(this::recuperarImagen).collect(Collectors.toList());		
 	}
 	
 	@Override
@@ -78,9 +72,44 @@ public class ServicioFotos implements IServicioFotos {
 		List<Long> archivos = Arrays.asList(dniFrontal, dniTrasero, certDelitos);
 		
 		for (Long long1 : archivos) {
-			repositorioImagen.deleteById(long1);
+			if (repositorioImagen.existsById(long1))
+				repositorioImagen.deleteById(long1);
 		}		
 
 	}
 	
+	private boolean esArchivoValido(MultipartFile file) {
+		return file != null && !file.isEmpty();
+	}
+
+	private Imagen crearImagen(MultipartFile file) {
+		Imagen img = null;
+		try {
+			byte[] originalBytes = file.getBytes();
+	        byte[] webpBytes = webPConverter.convertirAWebP(originalBytes);
+			img = new Imagen(file.getOriginalFilename(), webpBytes);
+		} catch (IOException e) {
+			e.printStackTrace();
+		}
+		return img;
+	}
+	
+	private Imagen recuperarImagen(Imagen i) {
+		Imagen img = null;
+		try {
+			byte[] originalBytes = i.getContenido();
+			
+			BufferedImage bufferedImage = webPConverter.descomprimirWebP(originalBytes);
+			
+			ByteArrayOutputStream byteArrayOutputStream = new ByteArrayOutputStream();
+			ImageIO.write(bufferedImage, i.getNombre().substring(i.getNombre().lastIndexOf('.') + 1).toLowerCase(), byteArrayOutputStream); 
+			byte[] imageBytes = byteArrayOutputStream.toByteArray();
+			
+			img = new Imagen(i.getNombre(), imageBytes);  
+			
+		} catch (IOException e) {
+			e.printStackTrace();
+		}
+		return img;
+}
 }
